fix wither prot flag not using correct blockpos; separated flags for wither (like in forge)
